#1213: now making sure that migrated folders don't already exist before migrating DTV folders (which can happen when you have *both* DTV and Democracy folders if you run/test multiple versions locally).
git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@1213 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 8a0928276459d4f7e0b0b4946be2ab317860b8c8

+++ tv/platform/osx/platformcfg.py
@@ -84,12 +84,18 @@ def migrateToDemocracy():
     # Migrate Movies and Support folders
 
     if os.path.exists(oldMoviesFolder):
-        print "DTV: Migrating movies folder to %s" % newMoviesFolder
-        os.rename(oldMoviesFolder, newMoviesFolder)
+        if not os.path.exists(newMoviesFolder):
+            print "DTV: Migrating movies folder to %s" % newMoviesFolder
+            os.rename(oldMoviesFolder, newMoviesFolder)
+        else:
+            print "DTV: WARNING! Both DTV and Democracy movies folder exist."
 
     if os.path.exists(oldSupportFolder):
-        print "DTV: Migrating support folder to %s" % newSupportFolder
-        os.rename(oldSupportFolder, newSupportFolder)
+        if not os.path.exists(newSupportFolder):
+            print "DTV: Migrating support folder to %s" % newSupportFolder
+            os.rename(oldSupportFolder, newSupportFolder)
+        else:
+            print "DTV: WARNING! Both DTV and Democracy support folder exist."
 
 def ensureMigratedMoviePath(pathname):
     if pathname.startswith(oldMoviesFolder):
